LangGraph-TagExtraction-proj2/requirements.txt
langchain~=0.3.25
langchain_groq~=0.3.2
langchain_openai~=0.3.23
langchain_community~=0.3.24
python-dotenv~=1.1.0
pyyaml~=6.0.2
langgraph~=0.4.8
pyjokes~=0.8.3
black~=25.1.0
spacy~=3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl
tqdm~=4.67.1
mcp~=1.9.4
langchain_mcp_adapters~=0.1.7
aiohttp>=3.11.14,<4.0.0
pygithub~=2.6.1
langchain_tavily~=0.2.5
selenium~=4.34.0
unstructured~=0.18.1